requirements.txt
streamlit==1.25.0
rembg==2.0.50
Pillow
streamlit-extras==0.3.0
